rename deepqmc.tools to deepqmc.extra
