+++ setup.cfg
@@ -17,7 +17,7 @@ combine_as_imports = true
 
 [tool:pytest]
 filterwarnings =
-    ignore:Code in deepqmc.tools:UserWarning
+    ignore:Code in deepqmc.extra:UserWarning
 norecursedirs = vendor
 
 [coverage:run]
@@ -30,7 +30,7 @@ omit =
     */deepqmc/torchext/sloglindet.py
     */deepqmc/torchext/bdet.py
     */deepqmc/utils.py
-    */deepqmc/tools/**.py
+    */deepqmc/extra/**.py
 
 [pydocstyle]
 add-ignore = D100,D102,D104,D105,D107,D202
