+++ src/deepqmc/extra/__init__.py
@@ -1,0 +1,3 @@
+from warnings import warn
+
+warn('Code in deepqmc.extra is not regularly tested and properly maintained')
