+++ src/deepqmc/extra/analysis.py
