+++ src/deepqmc/extra/grad.py
