+++ src/deepqmc/extra/memory.py
