+++ src/deepqmc/extra/plot.py
