--- src/deepqmc/tools/__init__.py
@@ -1,3 +1,0 @@
-from warnings import warn
-
-warn('Code in deepqmc.tools is not regularly tested and properly maintained')
